# src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Correos - Node.js",
    "description": "Documentación de la API Correos",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/paquetes/": {
      "post": {
        "tags": [
          "Paquetes"
        ],
        "description": "Registrar un nuevo paquete",
        "parameters": [
          {
            "name": "0",
            "in": "body",
            "description": "Datos del paquete",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "dimensiones": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        }
                      }
                    },
                    "remitente": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "object"
                        }
                      }
                    },
                    "direccion_remitente": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "object"
                        }
                      }
                    },
                    "destinatario": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "object"
                        }
                      }
                    },
                    "direccion_destinatario": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "object"
                        }
                      }
                    },
                    "peso": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "number"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Paquete registrado correctamente"
          },
          "400": {
            "description": "Datos inválidos"
          },
          "500": {
            "description": "Error en el servidor"
          }
        }
      }
    },
    "/paquetes/{id}": {
      "get": {
        "tags": [
          "Paquetes"
        ],
        "description": "Obtener paquete por ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Paquete encontrado"
          },
          "404": {
            "description": "No encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        }
      }
    },
    "/paquetes/paquetes/": {
      "post": {
        "tags": [
          "Paquetes"
        ],
        "description": "Obtener paquetes por usuario",
        "parameters": [
          {
            "name": "id",
            "in": "body",
            "required": true,
            "description": "ID del usuario",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Paquetes obtenidos"
          },
          "401": {
            "description": "No autorizado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/paquetes/gencode/{id}": {
      "get": {
        "tags": [
          "Paquetes"
        ],
        "description": "Generar código de barras en PNG",
        "produces": [
          "image/png"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Código de barras generado"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "No encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/paquetes/precio/{tamanio}/{peso}": {
      "get": {
        "tags": [
          "Paquetes"
        ],
        "description": "Calcular precio de envío",
        "parameters": [
          {
            "name": "tamaño",
            "in": "path",
            "required": true,
            "description": "Tamaño del paquete",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "peso",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "tamanio",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Precio calculado"
          },
          "400": {
            "description": "Parámetros inválidos"
          },
          "500": {
            "description": "Error en el servidor"
          }
        }
      }
    },
...
    "/envios/tracking/{usuario}": {
      "post": {
        "tags": [
          "Envios"
        ],
        "description": "Endpoint para registrar un nuevo envio",
        "parameters": [
          {
            "name": "0",
            "in": "body",
            "description": "Datos del envio",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        }
                      }
                    },
                    "usuario": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        }
                      }
                    },
                    "tipo": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          {
            "name": "usuario",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Envio registrado correctamente"
          },
          "400": {
            "description": "Datos inválidos"
          },
          "500": {
            "description": "Error en el servidor"
          }
        }
      }
    },
    "/envios/tracking/{id}": {
      "get": {
        "tags": [
          "Envios"
        ],
        "description": "Endpoint para obtener el estado de un envio",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Envio encontrado"
          },
          "404": {
            "description": "No encontrado"
          },
          "500": {
            "description": "Error en el servidor"
          }
        }
      }
    },
...
  }
}